re-introduced app user for running wiki
Controlling the uid and gid of 'app' user resolves the earlier problem
of writing to the /home/app/.wiki folder.  Sadly, it is implicitly
coupled to values chosen by docker-machine which are outside of our
control.  So these are vulnerable to environmental differences and the
resulting error messages do not lead easily to this point of coupling.

But I prefer that risk to the other risks associated with running the
app as root.